requirements.txt
fastapi==0.104.1
pydantic==2.4.2
python-dotenv==1.0.0
SQLAlchemy==1.3.6
SQLAlchemy_Utils==0.41.1
uvicorn==0.22.0